feat(community): add noice-nvim, disable lsp-signature-nvim
Not sure if this “highly experimental” toolkit is ready for GA, but many
plugins integrate with it, inc-rename.nvim implicitly depends on it,
and it does look noice.

It provides its own implementation of the LSP
`textDocument/signatureHelp` interface, so disable lsp_signature.nvim
in favor of noice.nvim's implementation.